+++ backend/src/main/java/com/studyhelper/entity/Authorities.java
@@ -1,5 +1,7 @@
 package com.studyhelper.entity;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -22,7 +24,7 @@ public class Authorities {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private int id;
+    private BigDecimal id;
 
     private String authority;
