+++ backend/src/main/java/com/studyhelper/service/CourseService.java
@@ -2,6 +2,7 @@ package com.studyhelper.service;
 
 import com.studyhelper.entity.Course;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CourseService {
@@ -9,9 +10,9 @@ public interface CourseService {
 
     void add(Course course);
 
-    Course getById(int id);
+    Course getById(BigDecimal id);
 
     Course getByName(String name);
 
-    void remove(int id);
+    void remove(BigDecimal id);
 }
