+++ backend/src/main/java/com/studyhelper/service/UserService.java
@@ -2,6 +2,7 @@ package com.studyhelper.service;
 
 import com.studyhelper.entity.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserService {
@@ -9,9 +10,9 @@ public interface UserService {
 
     void add(User user);
 
-    User getById(int id);
+    User getById(BigDecimal id);
 
     User getByUsername(String username);
 
-    void remove(int user);
+    void remove(BigDecimal user);
 }
